Build aggregation table

Build a table to quickly aggregate from pixels to polygons

Usage

build_aggregation_table(polygons, id_raster, polygon_id_field, verbose = FALSE)

Arguments

polygons

terra::SpatVector object. Should contain a unique ID field.

id_raster

terra::SpatRaster object. ID raster created by build_id_raster() for the polygons object. Should have the same CRS as polygons and completely cover it.

polygon_id_field

(character(1)) Unique identifier field in polygons.

verbose

(logical(1), default FALSE) Show progress for building aggregation rows for each polygon?

Value

data.table with fields:

polygon_id: Unique polygon identifier

pixel_id: unique pixel ID from the ID raster

masked_pixel_id: Index counting only non-NA pixels from the ID raster

area_fraction: fraction of the pixel area falling within this polygon

Merged fields from the table of polygons

See also