Remove global indent for logging

Change a global option that controls indentation for MBG package logging

Usage

logging_drop_indent()

Details

Decreases the MbgLoggingIndent option, with default of zero if not set